Reconfiguring, cleaning up the model
-Added W_anode_chem.cti to repo tracking.
-Changed file names to remove 'prelim'
-Changed main file name to `sei_1d_model.py`
-Made file names all lowercase
-Began cleaning up `sei_1d_model.py`
-Allow user to specify number of cycles

Model currently initializes a solution vector and creates a lookup
table for interpolating the working electrode electric potential
as a function of time.

The program integrates over the specified number of cycles, but
currently assumes that the system is always at steady state (i.e.
all residuals/time derivatives are set equal to zero).